analysis.os: New kernel memory usage view with Unit tests
This view keeps track of page allocation/deallocation events in the kernel.
It consists of a plot and a menu for thread selection. For a given time range,
the menu only lists threads in which there was activity. Values in the plot
are relative to the beginning of the selected trace.

Change-Id: Ie5605b7211828fd01accb920a5a735500fcae4b2
Signed-off-by: Najib Arbaoui <arbaouinajib@gmail.com>
Signed-off-by: Samuel Gagnon <samuel.gagnon92@gmail.com>
Reviewed-on: https://git.eclipse.org/r/65957
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>